decimal price

+++ UI/Controllers/HomeController.cs
@@ -44,7 +44,7 @@ namespace UI.Controllers
 								  Quantity = basket.Quantity,
 								  Price = basket.Price,
 								  ImagePath = product.ImagePath,
-								  TotalPrice = basket.Price * basket.Quantity,
+								  TotalPrice = (int)(basket.Price * basket.Quantity),
 								  Status = basket.Status
 							  })
 				  .Where(x => x.UserId == userId && x.Status==true)
